Create tagging capability
This allows users that are logged in and own a video to tag their video.
This also allows admins to add/edit tags for videos. If a user is on the
site and not logged in (Guest), they will be able to see the tags, but
not edit them or add to them.